mcp-sqlServer/requirements.txt
httpx
mcp[cli]>=1.5.0
# pyodbc  # Commented out - not needed for REST API approach
azure-identity